style-transfer/requirements.txt
aiohttp==3.1.0
tensorflow==1.8.0
scipy==1.1.0
aiobotocore==0.8.0
aiopg==0.14.0
psycopg2-binary==2.7.4
pillow==5.1.0
aiojobs==0.2.1
